+++ packages/sapa/src/Hook.js
@@ -338,9 +338,9 @@ export function useComponentRender(
 
   return component.useSubscribe(
     name,
-    () => {
+    (...args) => {
       if (isFunction(options.checkFunction)) {
-        if (options.checkFunction()) {
+        if (options.checkFunction(...args)) {
           useRender(component);
         }
       } else {